Move tests to same package as tested classes.
